RISC-V: Port FDT from riscv-pk (priv-1.10)

+++ src/plat/spike/machine/hardware.c
@@ -133,7 +133,7 @@ static inline uint64_t get_cycles(void)
         "rdtimeh %0\n"
         "rdtime  %1\n"
         : "=r" (nH), "=r" (nL));
-    return ((uint64_t) (nH << 32)) | (nL);
+    return ((uint64_t) ((uint64_t) nH << 32)) | (nL);
 }
 #else
 {
